# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: |
    RAWE Dog API — generates tailored job-application kits from a user's knowledge
    library and job postings, with live job discovery via TheirStack.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generate:
    post:
      operationId: generate
      tags: [generate]
      summary: Generate an application kit
      description: |
        Runs the kit generation pipeline. Four modes:

        - **stream** — Server-Sent Events; returns `text/event-stream`. Events:
          `status`, `pass1`, `draft`, `qa`, `repair`, `done`, `error`, `close`.
        - **pass1** — Selection only (JSON). Returns experience selection for
          user review before committing to drafting.
        - **pass2** — Drafting from an existing selection (JSON). Requires
          `selection` field.
        - **full** — One-shot full pipeline (JSON). Default if `mode` is omitted.

        When `postingId` is supplied the server resolves the posting's canonical
        brief from the cache (token-efficient) instead of re-sending the raw
        description. On successful kit generation the posting's status is
        automatically set to `kit_generated`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateRequest"
      responses:
        "200":
          description: Kit result (JSON modes) or SSE stream (mode=stream)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerateResult"
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream of GenerateProgressEvent objects
        "400":
          description: Missing required input or library not ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: postingId not found in cache
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Postings ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
        error:
          type: string
      required: [error]
...
    GenerateRequest:
      type: object
      properties:
        jobPosting:
          type: string
          description: >
            Raw job posting text. May be omitted when `postingId` is supplied
            and the posting has a usable canonical brief.
        postingId:
          type: string
          description: >
            ID of a stored posting. The server resolves the canonical brief from
            the cache so the raw description is never re-sent to the LLM.
        company:
          type: string
        targetTitle:
          type: string
        notes:
          type: string
          description: Extra user context appended to the generation prompt
        overrideLeads:
          type: array
          items:
            type: string
          description: Catalog IDs (E1…) to force-select, bypassing Pass 1
        skipPass1:
          type: boolean
        mode:
          type: string
          enum: [full, pass1, pass2, stream]
          default: full
        selection:
          $ref: "#/components/schemas/Pass1Selection"
    Pass1Selection:
      type: object
      description: Experience selection produced by Pass 1 (mode=pass1).
      properties:
        leadExperienceIds:
          type: array
          items:
            type: string
          description: Catalog IDs of selected lead experiences
        rationale:
          type: string
        targetTitle:
          type: string
        company:
          type: string
      required: [leadExperienceIds]
    ApplicationKit:
      type: object
      properties:
        meta:
          type: object
          properties:
            targetTitle:
              type: string
            company:
              type: string
            leadExperiences:
              type: array
              items:
                type: string
            rationale:
              type: string
            sourcesUsed:
              type: array
              items:
                type: string
          required: [targetTitle, company, leadExperiences, rationale, sourcesUsed]
        resumeMarkdown:
          type: string
        coverLetterMarkdown:
          type: string
        alignmentNotesMarkdown:
          type: string
        starPrepMarkdown:
          type: string
      required:
        [meta, resumeMarkdown, coverLetterMarkdown, alignmentNotesMarkdown, starPrepMarkdown]
    GenerateResult:
      type: object
      description: >
        JSON response for mode=full, mode=pass1, and mode=pass2.
        For mode=stream the response is a text/event-stream — this schema does
        not apply.
      properties:
        ok:
          type: boolean
        kit:
          $ref: "#/components/schemas/ApplicationKit"
        selection:
          $ref: "#/components/schemas/Pass1Selection"
        qaReport:
          type: object
          description: Verification findings (present when verification ran)
          properties:
            findings:
              type: array
              items:
                type: string
            keywordCoverage:
              type: array
              items:
                type: string
            repaired:
              type: array
              items:
                type: string
      required: [ok]

    # ── Postings ──────────────────────────────────────────────────────────────
...